backend/python/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0

# Socket.io
python-socketio==5.11.0
aiohttp==3.11.0

# Redis (Phase 2)
redis==5.2.0

# PostgreSQL (Phase 3)
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0

# Authentication (Phase 3)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17

# HTTP client
httpx==0.28.0

# Export (Phase 7)
pillow==11.0.0
reportlab==4.2.0

# Push Notifications (Phase 7)
firebase-admin==6.6.0

# Development
pytest==8.3.0
pytest-asyncio==0.24.0
